backend/requirements.txt
uvicorn
fastapi
sqlalchemy
python-dotenv
pydantic
psycopg2-binary
bcrypt